refactor: Extract in view
