+++ Mail/Views/Switch User/ContactCardButton.swift
@@ -1,0 +1,51 @@
+/*
+ Infomaniak Mail - iOS App
+ Copyright (C) 2025 Infomaniak Network SA
+
+ This program is free software: you can redistribute it and/or modify
+ it under the terms of the GNU General Public License as published by
+ the Free Software Foundation, either version 3 of the License, or
+ (at your option) any later version.
+
+ This program is distributed in the hope that it will be useful,
+ but WITHOUT ANY WARRANTY; without even the implied warranty of
+ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ GNU General Public License for more details.
+
+ You should have received a copy of the GNU General Public License
+ along with this program. If not, see <http://www.gnu.org/licenses/>.
+ */
+
+import DesignSystem
+import Foundation
+import MailResources
+import SwiftUI
+
+struct ContactCardButton: View {
+    @Binding var isShowingContactCard: Bool
+
+    @State private var hasFlipped = false
+
+    private let baseColor = UserDefaults.shared.accentColor
+
+    var body: some View {
+        Button {
+            isShowingContactCard = true
+        } label: {
+            MailResourcesAsset.qrCode.swiftUIImage
+                .iconSize(.medium)
+                .scaleEffect(x: hasFlipped ? -1 : 1, y: 1)
+                .foregroundStyle(.tint)
+                .padding(IKPadding.mini)
+                .background(
+                    RoundedRectangle(cornerRadius: IKRadius.medium)
+                        .foregroundStyle(baseColor.primary.swiftUIColor.opacity(0.1))
+                )
+                .onAppear {
+                    withAnimation(.default.delay(0.25)) {
+                        hasFlipped.toggle()
+                    }
+                }
+        }
+    }
+}
